Update fcpdf
Apr 17, 2020 version

01. Modified fcpdf so that it self-determines where it
is running from; and can therefore attach the header
and footer from that directory. This allows you to run
fcpdf from any other directory; as long as fcpdf is
either on your path; or if you specify the full path
fcpdf is in.

For example $ fcpdf md_file -OR- ~/scripts/fcpdf md_file

02. check to see if user supplied a ".md" on the
markdown file and do not auto-attach ".md" if they did so.
 
03.  By default a .pdf of the same name as the markdown
input file  is created. This mod allows the user to override this.
Allow user to add a second parameter, to create a
differently-named output .pdf.

usage:  $ fcpdf md_file  [full][path][to]pdf_file
.pdf will be appended to the output, if it is not supplied.